chore(video): opacity removed
